Refactored panels into subclasses, added loadable and saveable interface, added abstractToolBarPanel class -> take care of loading and saving when components are editing
