+++ Server/src/main/java/enums/FileType.java
@@ -1,4 +1,4 @@
-package Enums;
+package enums;
 
 public enum FileType {
     FILE(0),
