+++ Server/src/main/java/interfaces/Loadable.java
@@ -1,0 +1,5 @@
+package interfaces;
+
+public interface Loadable {
+    void load();
+}
