+++ Server/src/main/java/interfaces/Saveable.java
@@ -1,0 +1,6 @@
+package interfaces;
+
+public interface Saveable {
+
+    void save();
+}
