+++ Server/src/main/java/model/CFile.java
@@ -1,7 +1,7 @@
-package Model;
+package model;
 
 
-import Enums.FileType;
+import enums.FileType;
 import java.io.File;
 
 public class CFile extends Filesystem {
