+++ Server/src/main/java/view/CustomToolbar.java
@@ -1,4 +1,4 @@
-package View;
+package view;
 
 import javax.swing.*;
 import java.util.ArrayList;
