+++ Server/src/main/java/view/ToolbarButton.java
@@ -1,4 +1,4 @@
-package View;
+package view;
 
 import javax.swing.*;
