+++ Server/src/main/java/view/ToolbarButtonBuilder.java
@@ -1,4 +1,4 @@
-package View;
+package view;
 
 import javax.swing.*;
