+++ config.properties
@@ -1,11 +1,12 @@
 #--- HomeStream User Configuration ---
-#Sun Aug 02 14:46:08 PDT 2020
+#Sun Aug 02 19:32:34 PDT 2020
 start_on_login=false
-movie_folder=null
-password=admin
+movie_folder=/Users/jonathanhirsch/Movies
+password=hi5ethfkdfh
 local_port=3004
-use_different_remote_port=true
-remote_access_enabled=false
+remote_port=3004
+use_different_remote_port=false
+remote_access_enabled=true
 require_authentication=false
 auto_launch_server=false
-username=admin
+username=lololol test
